# ENH: faster circumcenter calculation in spatial.SphericalVoronoi