--- a/scipy/spatial/_spherical_voronoi.py
+++ b/scipy/spatial/_spherical_voronoi.py
@@ -11,12 +11,14 @@ Spherical Voronoi Code
 # Distributed under the same BSD license as SciPy.
 #
 
+import warnings
 import numpy as np
 import scipy
 from . import _voronoi
 from scipy.spatial import cKDTree
 
 __all__ = ['SphericalVoronoi']
+
 
 def sphere_check(points, radius, center):
     """ Determines distance of generators from theoretical sphere
@@ -28,65 +30,6 @@ def sphere_check(points, radius, center):
                             ((points[...,2] - center[2]) ** 2))
     max_discrepancy = (np.sqrt(actual_squared_radii) - radius).max()
     return abs(max_discrepancy)
-
-def calc_circumcenters(tetrahedrons):
-    """ Calculates the cirumcenters of the circumspheres of tetrahedrons.
-
-    An implementation based on
-    http://mathworld.wolfram.com/Circumsphere.html
-
-    Parameters
-    ----------
-    tetrahedrons : an array of shape (N, 4, 3)
-        consisting of N tetrahedrons defined by 4 points in 3D
-
-    Returns
-    ----------
-    circumcenters : an array of shape (N, 3)
-        consisting of the N circumcenters of the tetrahedrons in 3D
-
-    """
-
-    num = tetrahedrons.shape[0]
-    a = np.concatenate((tetrahedrons, np.ones((num, 4, 1))), axis=2)
-
-    sums = np.sum(tetrahedrons ** 2, axis=2)
-    d = np.concatenate((sums[:, :, np.newaxis], a), axis=2)
-
-    dx = np.delete(d, 1, axis=2)
-    dy = np.delete(d, 2, axis=2)
-    dz = np.delete(d, 3, axis=2)
-
-    dx = np.linalg.det(dx)
-    dy = -np.linalg.det(dy)
-    dz = np.linalg.det(dz)
-    a = np.linalg.det(a)
-
-    nominator = np.vstack((dx, dy, dz))
-    denominator = 2*a
-    return (nominator / denominator).T
-
-
-def project_to_sphere(points, center, radius):
-    """
-    Projects the elements of points onto the sphere defined
-    by center and radius.
-
-    Parameters
-    ----------
-    points : array of floats of shape (npoints, ndim)
-             consisting of the points in a space of dimension ndim
-    center : array of floats of shape (ndim,)
-            the center of the sphere to project on
-    radius : float
-            the radius of the sphere to project on
-
-    returns: array of floats of shape (npoints, ndim)
-            the points projected onto the sphere
-    """
-
-    lengths = scipy.spatial.distance.cdist(points, np.array([center]))
-    return (points - center) / lengths * radius + center
 
 
 class SphericalVoronoi:
@@ -114,10 +57,8 @@ class SphericalVoronoi:
         the points in 3D to generate the Voronoi diagram from
     radius : double
         radius of the sphere
-        Default: None (forces estimation, which is less precise)
     center : double array of shape (3,)
         center of the sphere
-        Default: None (assumes sphere is centered at origin)
     vertices : double array of shape (nvertices, 3)
         Voronoi vertices corresponding to points
     regions : list of list of integers of shape (npoints, _ )
@@ -135,35 +76,19 @@ class SphericalVoronoi:
     The spherical Voronoi diagram algorithm proceeds as follows. The Convex
     Hull of the input points (generators) is calculated, and is equivalent to
     their Delaunay triangulation on the surface of the sphere [Caroli]_.
-    A 3D Delaunay tetrahedralization is obtained by including the origin of
-    the coordinate system as the fourth vertex of each simplex of the Convex
-    Hull. The circumcenters of all tetrahedra in the system are calculated and
-    projected to the surface of the sphere, producing the Voronoi vertices.
-    The Delaunay tetrahedralization neighbour information is then used to
+    The Convex Hull neighbour information is then used to
     order the Voronoi region vertices around each generator. The latter
     approach is substantially less sensitive to floating point issues than
     angle-based methods of Voronoi region vertex sorting.
 
-    The surface area of spherical polygons is calculated by decomposing them
-    into triangles and using L'Huilier's Theorem to calculate the spherical
-    excess of each triangle [Weisstein]_. The sum of the spherical excesses is
-    multiplied by the square of the sphere radius to obtain the surface area
-    of the spherical polygon. For nearly-degenerate spherical polygons an area
-    of approximately 0 is returned by default, rather than attempting the
-    unstable calculation.
-
     Empirical assessment of spherical Voronoi algorithm performance suggests
     quadratic time complexity (loglinear is optimal, but algorithms are more
-    challenging to implement). The reconstitution of the surface area of the
-    sphere, measured as the sum of the surface areas of all Voronoi regions,
-    is closest to 100 % for larger (>> 10) numbers of generators.
+    challenging to implement).
 
     References
     ----------
     .. [Caroli] Caroli et al. Robust and Efficient Delaunay triangulations of
                 points on or close to a sphere. Research Report RR-7004, 2009.
-    .. [Weisstein] "L'Huilier's Theorem." From MathWorld -- A Wolfram Web
-                Resource. http://mathworld.wolfram.com/LHuiliersTheorem.html
 
     See Also
     --------
@@ -216,16 +141,26 @@ class SphericalVoronoi:
 
     """
 
-    def __init__(self, points, radius=None, center=None, threshold=1e-06):
+    def __init__(self, points, radius=1, center=(0, 0, 0), threshold=1e-06):
+        if radius is None:
+            radius = 1.
+            warnings.warn('`radius` is `None`. '
+                          'This will raise an error in a future version. '
+                          'Please provide a floating point number '
+                          '(i.e. `radius=1`).',
+                          DeprecationWarning)
+
+        if center is None:
+            center = (0, 0, 0)
+            warnings.warn('`center` is `None`. '
+                          'This will raise an error in a future version. '
+                          'Please provide a coordinate '
+                          '(i.e. `center=(0, 0, 0)`)',
+                          DeprecationWarning)
+
         self.points = points
-        if np.any(center):
-            self.center = center
-        else:
-            self.center = np.zeros(3)
-        if radius:
-            self.radius = radius
-        else:
-            self.radius = 1
+        self.radius = radius
+        self.center = np.array(center)
 
         if cKDTree(self.points).query_pairs(threshold * self.radius):
             raise ValueError("Duplicate generators present.")
@@ -250,32 +185,15 @@ class SphericalVoronoi:
         Tyler Reddy, Ross Hemsley and Nikolai Nowaczyk
         """
 
-        # perform 3D Delaunay triangulation on data set
-        # (here ConvexHull can also be used, and is faster)
+        # get Convex Hull
         self._tri = scipy.spatial.ConvexHull(self.points)
 
-        # add the center to each of the simplices in tri to get the same
-        # tetrahedrons we'd have gotten from Delaunay tetrahedralization
-        # tetrahedrons will have shape: (2N-4, 4, 3)
-        tetrahedrons = self._tri.points[self._tri.simplices]
-        tetrahedrons = np.insert(
-            tetrahedrons,
-            3,
-            np.array([self.center]),
-            axis=1
-        )
+        # triangles will have shape: (2N-4, 3, 3)
+        triangles = self._tri.points[self._tri.simplices]
 
-        # produce circumcenters of tetrahedrons from 3D Delaunay
+        # get circumcenters of Convex Hull triangles from facet equations
         # circumcenters will have shape: (2N-4, 3)
-        circumcenters = calc_circumcenters(tetrahedrons)
-
-        # project tetrahedron circumcenters to the surface of the sphere
-        # self.vertices will have shape: (2N-4, 3)
-        self.vertices = project_to_sphere(
-            circumcenters,
-            self.center,
-            self.radius
-        )
+        self.vertices = self.radius * self._tri.equations[:, :-1] + self.center
 
         # calculate regions from triangulation
         # simplex_indices will have shape: (2N-4,)
@@ -308,14 +226,13 @@ class SphericalVoronoi:
 
         This is done as follows: Recall that the n-th region in regions
         surrounds the n-th generator in points and that the k-th
-        Voronoi vertex in vertices is the projected circumcenter of the
-        tetrahedron obtained by the k-th triangle in _tri.simplices (and the
-        origin). For each region n, we choose the first triangle (=Voronoi
-        vertex) in _tri.simplices and a vertex of that triangle not equal to
-        the center n. These determine a unique neighbor of that triangle,
-        which is then chosen as the second triangle. The second triangle
-        will have a unique vertex not equal to the current vertex or the
-        center. This determines a unique neighbor of the second triangle,
+        Voronoi vertex in vertices is the circumcenter of the k-th triangle
+        in _tri.simplices.  For each region n, we choose the first triangle
+        (=Voronoi vertex) in _tri.simplices and a vertex of that triangle
+        not equal to the center n. These determine a unique neighbor of that
+        triangle, which is then chosen as the second triangle. The second
+        triangle will have a unique vertex not equal to the current vertex or
+        the center. This determines a unique neighbor of the second triangle,
         which is then chosen as the third triangle and so forth. We proceed
         through all the triangles (=Voronoi vertices) belonging to the
         generator in points and obtain a sorted version of the vertices