<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
https://docs.scipy.org/doc/numpy/dev/gitwash/development_workflow.html#writing-the-commit-message

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
<!--Example: Closes gh-WXYZ.-->
Closes #9386, #8859, #10484

#### What does this implement/fix?
Implements a faster algorithm for calculation of circumcenters, with better numerical stability.
Benchmarks:
```
============ ============= =============
 num_points      master       this PR
------------ ------------- -------------
     10         494±4μs      410±5μs   
    100         1.32±0ms     955±3μs   
    1000      10.2±0.08ms   6.85±0.1ms 
    5000      53.6±0.07ms   35.2±0.6ms 
   10000       112±0.6ms     76.4±1ms  
============ ============= =============
```

-Resolves circumcenter signs properly, which means points in a single hemisphere are correctly handled.

-Updated docstrings

-Changed default parameter values to their actual values, rather than `None`.

#### Additional information
<!--Any additional information you think is important.-->
-A remaining edge case is input data consisting of a point set on a great circle.  If the points are equi-spaced, the correct output should resemble the segments of an orange.  I can implement this in a later PR.

~~-The new reference is a paper of mine.  I thought it might be of interest if readers would like to understand the algorithm for calculating circumcenters.  If it is not appropriate I don't mind taking it out again.~~