@pmla 2019-07-28
After some investigating it turns out that we can get the information we need directly from Qhull.  The correct circumcenters are stored in the normal component of the facet equations.  The results are identical to my original solution in this PR, but the current code is considerably simpler.

I have updated the benchmarks in my previous comment.