@rtshort 2019-08-18
Thanks for pinging me about this.  This does look way better.  I am very interested in getting the circumcenters directly from qhull and will spend some time looking at this code.  Some comments:

1. I am putting my own prejudices forth here, but compared to fixing the problem with points in one hemisphere is worth (to me) an API change.  I wonder how much usage this routine gets with the current limitation?  It would be better to have no API change of course. 

2.  There is another way of getting the correct circumcenters.  The outer hull, consisting of the facets not visible from the origin (see recent changes to qhull - Tyler helped me add the "good" element to ConvexHull) yield the points on one side of the sphere and the inner hull, the facets visible from the origin yield the points on the other side of the sphere.  I just spent the last several months, working weekend mornings for a couple of hours, building a SphericalVoronoi class that uses this concept.
  
3. Doing the circumcenters the way SphericalVoronoi does now takes the determinant of a 4x4 matrix, but the circumcenters may be found by finding the null space of a 3x3 matrix as well.  That is a significant reduction for a large problem.  On the other hand if qhull is already doing the work that is clearly the way to go.

4.  I think the spherical Voronoi diagram should look like the attached figure.  The ridges should not be hyperplanes in 3-space but should be great arcs on the sphere.  I am no geometer and could be wrong, but I think that is the case.
![ExampleSphericalVoronoi](https://user-images.githubusercontent.com/2798381/63225588-5063b180-c186-11e9-8c5c-dea34b4231c0.png)