@rtshort 2019-08-18
I looked at this code fairly carefully.  This is a substantial improvement over the current implementation.  It uses a non-obvious artifact of qhull, and I wish I had thought of it.  It also obviates the code I just spent the last year working on.  LMAO.  Oh well, it was a labor of love and a learning experience so not time lost.

The API change is really a no-op.  It just makes the defaults clear.  However if somebody is actually passing None for the radius or center their code will break.  I would recommend that we leave the API as pmla has it, but put the checks for None back in with a deprecation warning.   No API breakage and the future is brighter.

DB