@pmla 2019-08-18
Thanks @tylerjereddy and @rtshort for your comments.

> It also obviates the code I just spent the last year working on

If it is any consolation then I also threw out a lot of code I had written when I discovered this :)

> However if somebody is actually passing None for the radius or center their code will break. I would recommend that we leave the API as pmla has it, but put the checks for None back in with a deprecation warning

I really like this suggestion. I have implemented it.  @tylerjereddy what say you?