@rtshort 2019-08-19
I think we have the solution.  in spite of my personal mortification this is excellent.  i would like to continue the discussion about hyperplanes and such.  If we close this the conversation is lost. So,even though I think we have a great solution I would like to continue the conversation.  Of course I can open an issue and perhaps that is a better approach.

Also, I would like to spend some time on plma's last post.  Sometimes qhull is a little mysterious.  I can't do this every morning, but I would really like to understand this but it will take a while.  So, even though plma has really solved the problem, please let us take the time to understand.  I promise to spend the time I have, but it just isn't that much.

@plma (and I am still pretty github and markdown illiterate) I am impressed.

DB