@pmla 2019-08-19
@rtshort you have pinged a different user - off by one :
Here is an explanation of the bugfix.

How you determine the circumcentres doesn't really matter (the old algorithm is fine, if a bit slow).  What is important for single-hemisphere calculations is choosing the correct sign, i.e. which of the two possible unit normals.

To illustrate, here are some 2D examples of spherical Voronoi diagrams.  Points are shown in blue, the convex hull is a dotted blue line, circumcentres are green if correct and red if wrong.

Here is an easy case, where points lie in both hemispheres.  As in the old algorithm, the circumcentres are projected outwards, from the _origin_.  All circumcentres are correct.
![demo0](https://user-images.githubusercontent.com/5646170/63283502-5db09700-c27f-11e9-9b7f-10ebed9e8912.png)

If all points lie in a single hemisphere, projecting outwards from the origin produces a bad circumcentre (shown in red).  The reason for this is that the origin no longer lies in the interior of the convex hull.
![demo1](https://user-images.githubusercontent.com/5646170/63283501-5db09700-c27f-11e9-8dc2-479454716ec5.png)

To fix this problem, we project from an interior point of the set defined by the convex hull.  Any interior point will do.  Here I have picked the barycentre, which is guaranteed to lie in the interior.  The circumcentre is now projected in the correct direction:
![demo2](https://user-images.githubusercontent.com/5646170/63283500-5db09700-c27f-11e9-934f-857652fafc76.png)