@tylerjereddy 2019-08-19
@pmla Can you link to the paper you describe--we might want to cite that in the source code comment above the tricky line even if we don't mention in the public docstring. I'm also pretty sure that @rtshort has very similar conceptual diagrams in their explanation here: https://github.com/scipy/scipy/issues/9386#issuecomment-431070632