@pmla 2019-08-19
Here is the paper.  The proof is on page 4.
P M Larsen and S Schmidt.  Improved orientation sampling for indexing diffraction patterns of polycrystalline materials.
J. Appl. Cryst. (2017). 50, 1571-1582
https://doi.org/10.1107/S1600576717012882
https://arxiv.org/abs/1707.09045

> I'm also pretty sure that @rtshort has very similar conceptual diagrams in their explanation here: #9386 (comment)

Oh, right :)