@pmla 2019-08-19
> Ok, I'm going to merge this now, thanks--big improvement!

Great!  Thanks to all involved for the help with this PR.
I have closed #10484.  I recommend that #9386 and #8859 also be closed.

I have updated the 1.4.0 release notes wiki with this text:
> `scipy.spatial.SphericalVoronoi` now has linear memory complexity, improved performance, and supports single-hemisphere generators.


> (maybe ?) follow-up with a DOC PR that includes the circumcenter algorithm citation, possibly in the public docstring, if that is appropriate

I am biased, so I will not comment on this.  I am happy either way :)


Things I would like to do:
- PEP8 cleanup.  Is this allowed or does it just make changes harder to track?  I run `pep8` and `pyflakes` before pushing - at the moment I have to inspect the output to see if the warnings come from my changes or not.
- Add support for any numbers of dimensions, i.e. not just 3D.  My own need for `SphericalVoronoi` stems from analysis of points on a 4D sphere (the paper I linked uses is all about this).  The `spatial.Voronoi` and `spatial.ConvexHull` classes support arbitrary numbers of dimensions.  Apart from the `sort_vertices_of_regions` (which does not have a meaningful definition in higher dimensions) the current implementation can easily be adapted.